Keyframe indication in single NAL unit packet.

+++ src/net/java/sip/communicator/impl/neomedia/codec/video/h264/JNIEncoder.java
@@ -245,7 +245,7 @@ public class JNIEncoder
         // average bit rate
         FFmpeg.avcodeccontext_set_bit_rate(avcontext, bitRate);
         // so to be 1 in x264
-        FFmpeg.avcodeccontext_set_bit_rate_tolerance(avcontext, (int)(bitRate /
+        FFmpeg.avcodeccontext_set_bit_rate_tolerance(avcontext, (bitRate /
                 frameRate));
         FFmpeg.avcodeccontext_set_rc_max_rate(avcontext, bitRate);
         FFmpeg.avcodeccontext_set_sample_aspect_ratio(avcontext, 0, 0);
